requirements.txt
langchain
langchain-openai
langchain-anthropic
wikipedia
duckduckgo-search
python-dotenv
pydantic